ast: internalize the AST; drop the ruruby-parse dependency
Phase 2 (final) of the prism migration. monoruby no longer depends on
`ruruby-parse` at all — the AST types it consumed are now owned by
monoruby under `monoruby/src/ast/`.

- Move the AST + support modules into `monoruby/src/ast/`: `node`
  (Node/NodeKind/BlockInfo/FormalParam/ParamKind/ArgList/CaseBranch/
  CmpKind/BinOp/UnOp), `source_info` (Loc/SourceInfo/SourceInfoRef),
  `lvar_collector` (LvarCollector/…), `error` (ParseErr/ParseErrKind).
- Bring the remaining public AST types that lived in the parser crate's
  root/`parser.rs` into `ast.rs`: `Annot`, `LocalsContext`, `ParseResult`,
  `RescueEntry`, `NReal`.
- Prune the two parser-only constructors that reached into the lexer
  (`Node::new_string(RubyString)`, `ParseErr::from_lexer_err(LexerErr)`);
  the Prism lowerer builds these nodes directly. Repoint `crate::Loc` to
  `super::Loc` in lvar_collector.
- The Prism lowerer now names the AST through `crate::ast::*` instead of
  `ruruby_parse::*`.
- Add `console` (was a ruruby-parse dep, used by SourceInfo's error
  display). Drop `ruruby-parse` from the workspace members and
  `monoruby/Cargo.toml`, and delete the `ruruby-parse/` directory (its
  full parser/lexer was upstreamed to sisshiki1969/ruruby-parse).

cargo check (lib + bins + tests) is clean on x86_64 and
aarch64-unknown-linux-gnu. Under Ruby 4.0.2 the full lib suite (1644) and
tests/add_coverage (61) pass on x86_64.
