--- ruruby-parse/.vscode/launch.json
@@ -1,349 +1,0 @@
-{
-  // Use IntelliSense to learn about possible attributes.
-  // Hover to view descriptions of existing attributes.
-  // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
-  "version": "0.2.0",
-  "configurations": [
-    {
-      "type": "lldb",
-      "request": "launch",
-      "name": "Debug executable 'irm'",
-      "cargo": {
-        "args": [
-          "build",
-          "--bin=ruruby-parse",
-          "--package=ruruby-parse"
-        ],
-        "filter": {
-          "name": "ruruby-parse",
-          "kind": "bin"
-        }
-      },
-      "args": [
-        "../monoruby_p.rb"
-      ],
-      "cwd": "${workspaceFolder}"
-    },
-    {
-      "type": "lldb",
-      "request": "launch",
-      "name": "Debug unit tests in executable 'irm'",
-      "cargo": {
-        "args": [
-          "test",
-          "--no-run",
-          "--bin=irm",
-          "--package=monoruby"
-        ],
-        "filter": {
-          "name": "irm",
-          "kind": "bin"
-        }
-      },
-      "args": [],
-      "cwd": "${workspaceFolder}"
-    },
-    {
-      "type": "lldb",
-      "request": "launch",
-      "name": "Debug executable 'monoruby'",
-      "cargo": {
-        "args": [
-          "build",
-          "--bin=monoruby",
-          "--package=monoruby"
-        ],
-        "filter": {
-          "name": "monoruby",
-          "kind": "bin"
-        }
-      },
-      "args": [],
-      "cwd": "${workspaceFolder}"
-    },
-    {
-      "type": "lldb",
-      "request": "launch",
-      "name": "Debug unit tests in executable 'monoruby'",
-      "cargo": {
-        "args": [
-          "test",
-          "--no-run",
-          "--bin=monoruby",
-          "--package=monoruby"
-        ],
-        "filter": {
-          "name": "monoruby",
-          "kind": "bin"
-        }
-      },
-      "args": [],
-      "cwd": "${workspaceFolder}"
-    },
-    {
-      "type": "lldb",
-      "request": "launch",
-      "name": "Debug integration test 'arith'",
-      "cargo": {
-        "args": [
-          "test",
-          "--no-run",
-          "--test=arith",
-          "--package=monoruby"
-        ],
-        "filter": {
-          "name": "arith",
-          "kind": "test"
-        }
-      },
-      "args": [],
-      "cwd": "${workspaceFolder}"
-    },
-    {
-      "type": "lldb",
-      "request": "launch",
-      "name": "Debug integration test 'case'",
-      "cargo": {
-        "args": [
-          "test",
-          "--no-run",
-          "--test=case",
-          "--package=monoruby"
-        ],
-        "filter": {
-          "name": "case",
-          "kind": "test"
-        }
-      },
-      "args": [],
-      "cwd": "${workspaceFolder}"
-    },
-    {
-      "type": "lldb",
-      "request": "launch",
-      "name": "Debug integration test 'comparable'",
-      "cargo": {
-        "args": [
-          "test",
-          "--no-run",
-          "--test=comparable",
-          "--package=monoruby"
-        ],
-        "filter": {
-          "name": "comparable",
-          "kind": "test"
-        }
-      },
-      "args": [],
-      "cwd": "${workspaceFolder}"
-    },
-    {
-      "type": "lldb",
-      "request": "launch",
-      "name": "Debug integration test 'enumerable'",
-      "cargo": {
-        "args": [
-          "test",
-          "--no-run",
-          "--test=enumerable",
-          "--package=monoruby"
-        ],
-        "filter": {
-          "name": "enumerable",
-          "kind": "test"
-        }
-      },
-      "args": [],
-      "cwd": "${workspaceFolder}"
-    },
-    {
-      "type": "lldb",
-      "request": "launch",
-      "name": "Debug integration test 'literal'",
-      "cargo": {
-        "args": [
-          "test",
-          "--no-run",
-          "--test=literal",
-          "--package=monoruby"
-        ],
-        "filter": {
-          "name": "literal",
-          "kind": "test"
-        }
-      },
-      "args": [],
-      "cwd": "${workspaceFolder}"
-    },
-    {
-      "type": "lldb",
-      "request": "launch",
-      "name": "Debug integration test 'method_call'",
-      "cargo": {
-        "args": [
-          "test",
-          "--no-run",
-          "--test=method_call",
-          "--package=monoruby"
-        ],
-        "filter": {
-          "name": "method_call",
-          "kind": "test"
-        }
-      },
-      "args": [],
-      "cwd": "${workspaceFolder}"
-    },
-    {
-      "type": "lldb",
-      "request": "launch",
-      "name": "Debug integration test 'mul_assign'",
-      "cargo": {
-        "args": [
-          "test",
-          "--no-run",
-          "--test=mul_assign",
-          "--package=monoruby"
-        ],
-        "filter": {
-          "name": "mul_assign",
-          "kind": "test"
-        }
-      },
-      "args": [],
-      "cwd": "${workspaceFolder}"
-    },
-    {
-      "type": "lldb",
-      "request": "launch",
-      "name": "Debug integration test 'redefine'",
-      "cargo": {
-        "args": [
-          "test",
-          "--no-run",
-          "--test=redefine",
-          "--package=monoruby"
-        ],
-        "filter": {
-          "name": "redefine",
-          "kind": "test"
-        }
-      },
-      "args": [],
-      "cwd": "${workspaceFolder}"
-    },
-    {
-      "type": "lldb",
-      "request": "launch",
-      "name": "Debug integration test 'require'",
-      "cargo": {
-        "args": [
-          "test",
-          "--no-run",
-          "--test=require",
-          "--package=monoruby"
-        ],
-        "filter": {
-          "name": "require",
-          "kind": "test"
-        }
-      },
-      "args": [],
-      "cwd": "${workspaceFolder}"
-    },
-    {
-      "type": "lldb",
-      "request": "launch",
-      "name": "Debug integration test 'rescue'",
-      "cargo": {
-        "args": [
-          "test",
-          "--no-run",
-          "--test=rescue",
-          "--package=monoruby"
-        ],
-        "filter": {
-          "name": "rescue",
-          "kind": "test"
-        }
-      },
-      "args": [],
-      "cwd": "${workspaceFolder}"
-    },
-    {
-      "type": "lldb",
-      "request": "launch",
-      "name": "Debug integration test 'variables'",
-      "cargo": {
-        "args": [
-          "test",
-          "--no-run",
-          "--test=variables",
-          "--package=monoruby"
-        ],
-        "filter": {
-          "name": "variables",
-          "kind": "test"
-        }
-      },
-      "args": [],
-      "cwd": "${workspaceFolder}"
-    },
-    {
-      "type": "lldb",
-      "request": "launch",
-      "name": "Debug unit tests in library 'ruruby_parse'",
-      "cargo": {
-        "args": [
-          "test",
-          "--no-run",
-          "--lib",
-          "--package=ruruby-parse"
-        ],
-        "filter": {
-          "name": "ruruby_parse",
-          "kind": "lib"
-        }
-      },
-      "args": [],
-      "cwd": "${workspaceFolder}"
-    },
-    {
-      "type": "lldb",
-      "request": "launch",
-      "name": "Debug executable 'ruruby-parse'",
-      "cargo": {
-        "args": [
-          "build",
-          "--bin=ruruby-parse",
-          "--package=ruruby-parse"
-        ],
-        "filter": {
-          "name": "ruruby-parse",
-          "kind": "bin"
-        }
-      },
-      "args": [],
-      "cwd": "${workspaceFolder}"
-    },
-    {
-      "type": "lldb",
-      "request": "launch",
-      "name": "Debug unit tests in executable 'ruruby-parse'",
-      "cargo": {
-        "args": [
-          "test",
-          "--no-run",
-          "--bin=ruruby-parse",
-          "--package=ruruby-parse"
-        ],
-        "filter": {
-          "name": "ruruby-parse",
-          "kind": "bin"
-        }
-      },
-      "args": [],
-      "cwd": "${workspaceFolder}"
-    }
-  ]
-}
