--- ruruby-parse/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "ruruby-parse"
-version = "0.0.1"
-authors = ["monochrome <sisshiki@mac.com>"]
-edition = "2024"
-description = "Ruby parser written with Rust."
-license = "MIT OR Apache-2.0"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-fxhash = "0.2.1"
-once_cell = "1.21.3"
-enum-iterator = "2.1.0"
-num = "0.4"
-num-bigint = "0.4.2"
-console = "0.15.11"
-clap = { version = "4.5.35", features = ["derive", "cargo"] }
-# Use the pure-Rust rustls TLS backend instead of the default native-tls
-# (OpenSSL). reqwest is only used by this crate's own `main.rs` binary, and
-# avoiding the openssl-sys C dependency lets the workspace cross-compile to
-# aarch64 (for the aarch64 backend's qemu test environment) without a target
-# OpenSSL install.
-reqwest = { version = "0.12.15", default-features = false, features = [
-  "blocking",
-  "rustls-tls",
-] }
