--- ruruby-parse/src/parser.rs
@@ -1,1241 +1,0 @@
-use super::*;
-use num::BigInt;
-use std::path::PathBuf;
-
-mod arguments;
-mod define;
-mod expression;
-mod flow_control;
-mod lexer;
-mod literals;
-mod primary;
-pub(crate) use lexer::*;
-
-pub trait LocalsContext: Sized {
-    fn find_lvar(&self, id: &str) -> Option<usize>;
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct DummyContext;
-
-impl LocalsContext for DummyContext {
-    fn find_lvar(&self, _id: &str) -> Option<usize> {
-        None
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct Parser<'a, OuterContext: LocalsContext> {
-    lexer: Lexer<'a>,
-    path: PathBuf,
-    prev_loc: Loc,
-    /// scope stack.
-    scope: Vec<LvarScope>,
-    /// loop stack.
-    loop_stack: Vec<LoopKind>,
-    extern_context: Option<&'a OuterContext>,
-    /// this flag suppress accesory assignment. e.g. x=3
-    suppress_acc_assign: bool,
-    /// this flag suppress accesory multiple assignment. e.g. x = 2,3
-    suppress_mul_assign: bool,
-    /// this flag suppress parse do-end style block.
-    suppress_do_block: bool,
-    /// defined? mode: allow invalid break/next.
-    defined_mode: bool,
-    /// line number offset for eval (0-based).
-    line_offset: i64,
-}
-
-impl<'a> Parser<'a, DummyContext> {
-    pub fn parse_program(code: String, path: impl Into<PathBuf>) -> Result<ParseResult, ParseErr> {
-        let path = path.into();
-        let parse_ctx = LvarScope::new_eval(None);
-        parse(code, path, None::<&DummyContext>, parse_ctx, 0)
-    }
-}
-
-impl<'a, OuterContext: LocalsContext> Parser<'a, OuterContext> {
-    pub fn parse_program_eval(
-        code: String,
-        path: PathBuf,
-        extern_context: Option<&'a OuterContext>,
-        line_offset: i64,
-    ) -> Result<ParseResult, ParseErr> {
-        let parse_ctx = LvarScope::new_eval(None);
-        parse(code, path, extern_context, parse_ctx, line_offset)
-    }
-
-    pub fn parse_program_binding(
-        code: String,
-        path: PathBuf,
-        context: Option<LvarCollector>,
-        extern_context: Option<&OuterContext>,
-        line_offset: i64,
-    ) -> Result<ParseResult, ParseErr> {
-        let parse_ctx = LvarScope::new_block(context);
-        parse(code, path, extern_context, parse_ctx, line_offset)
-    }
-
-    fn parse(
-        code: &'a str,
-        path: PathBuf,
-        extern_context: Option<&'a OuterContext>,
-        scope: LvarScope,
-        line_offset: i64,
-    ) -> Result<(Node, LvarCollector, Token), LexerErr> {
-        let lexer = Lexer::new(code);
-        let mut parser = Parser {
-            lexer,
-            path,
-            prev_loc: Loc(0, 0),
-            scope: vec![scope],
-            loop_stack: vec![LoopKind::Top],
-            extern_context,
-            suppress_acc_assign: false,
-            suppress_mul_assign: false,
-            suppress_do_block: false,
-            defined_mode: false,
-            line_offset,
-        };
-        let node = parser.parse_comp_stmt()?;
-        let lvar = parser.scope.pop().unwrap().lvar;
-        let tok = parser.peek()?;
-        Ok((node, lvar, tok))
-    }
-
-    fn save_state(&self) -> (usize, usize) {
-        self.lexer.save_state()
-    }
-
-    fn restore_state(&mut self, state: (usize, usize)) {
-        self.lexer.restore_state(state);
-    }
-
-    fn scope_mut(&mut self) -> &mut LvarScope {
-        self.scope.last_mut().unwrap()
-    }
-
-    fn is_method_context(&self) -> bool {
-        for scope in self.scope.iter().rev() {
-            match scope.kind {
-                ScopeKind::Method => return true,
-                ScopeKind::Class | ScopeKind::Eval => return false,
-                ScopeKind::Block | ScopeKind::For => {}
-            }
-        }
-        false
-    }
-
-    /// Check whether parameter delegation exists or not in the method def of current context.
-    /// If not, return ParseErr.
-    fn check_forwarding(&self) -> Result<(), LexerErr> {
-        for ctx in self.scope.iter().rev() {
-            if ctx.kind == ScopeKind::Method {
-                if ctx.lvar.forwarding_param.is_some() {
-                    return Ok(());
-                } else {
-                    break;
-                }
-            }
-        }
-        Err(error_unexpected(self.prev_loc(), "Unexpected ..."))
-    }
-
-    /// If the `id` does not exist in the scope chain,
-    /// add `id` as a local variable in the current context.
-    fn add_local_var_if_new(&mut self, name: &str) -> usize {
-        if let Some(outer) = self.is_local_var(name) {
-            return outer;
-        } else {
-            for c in self.scope.iter_mut().rev() {
-                match c.kind {
-                    ScopeKind::For => {}
-                    _ => {
-                        c.lvar.insert(name);
-                        return 0;
-                    }
-                };
-            }
-        }
-        unreachable!()
-    }
-
-    /// Add the `id` as a new parameter in the current context.
-    /// If a parameter with the same name already exists, return error.
-    fn new_param(&mut self, name: &str, loc: Loc) -> Result<LvarId, LexerErr> {
-        match self.scope_mut().lvar.insert_new(name.to_string()) {
-            Some(lvar) => Ok(lvar),
-            None => Err(error_unexpected(
-                loc,
-                format!("Duplicated argument name `{}`.", name),
-            )),
-        }
-    }
-
-    fn add_kw_param(&mut self, lvar: LvarId) {
-        self.scope_mut().lvar.kw.push(lvar);
-    }
-
-    /// Add the `id` as a new parameter in the current context.
-    /// If a parameter with the same name already exists, return error.
-    fn new_kwrest_param(&mut self, name: String, loc: Loc) -> Result<(), LexerErr> {
-        if self.scope_mut().lvar.insert_kwrest_param(name).is_none() {
-            return Err(error_unexpected(loc, "Duplicated argument name."));
-        }
-        Ok(())
-    }
-
-    /// Add the `id` as a new block parameter in the current context.
-    /// If a parameter with the same name already exists, return error.
-    fn new_block_param(&mut self, name: String, loc: Loc) -> Result<(), LexerErr> {
-        if self.scope_mut().lvar.insert_block_param(name).is_none() {
-            return Err(error_unexpected(loc, "Duplicated argument name."));
-        }
-        Ok(())
-    }
-
-    /// Add the `id` as a new block parameter in the current context.
-    /// If a parameter with the same name already exists, return error.
-    fn new_forwarding_param(&mut self, loc: Loc) -> Result<(), LexerErr> {
-        if self.scope_mut().lvar.insert_delegate_param().is_none() {
-            return Err(error_unexpected(loc, "Duplicated argument name."));
-        }
-        Ok(())
-    }
-
-    /// Examine whether `id` exists in the scope chain.
-    /// If exiets, return true.
-    fn is_local_var(&mut self, id: &str) -> Option<usize> {
-        let mut outer = 0;
-        for c in self.scope.iter().rev() {
-            if c.lvar.table.get_lvarid(id).is_some() {
-                return Some(outer);
-            }
-            match c.kind {
-                ScopeKind::Block => outer += 1,
-                ScopeKind::For => {}
-                _ => break,
-            }
-        }
-        if let Some(a) = &self.extern_context {
-            return a.find_lvar(id);
-        }
-        None
-    }
-
-    /// Check whether numbered parameters exist in outer blocks.
-    fn check_outer_numbered_param(&mut self, loc: Loc) -> Result<(), LexerErr> {
-        //let mut outer = 0;
-        if let Some(loc) = self.scope.last_mut().unwrap().lvar.prohibit_numbered_param {
-            return Err(error_nested_numbered_param(loc));
-        }
-        for c in self.scope.iter_mut().rev().skip(1) {
-            if let Some(loc) = c.lvar.numbered_param {
-                return Err(error_nested_numbered_param(loc));
-            }
-            let prohibit = &mut c.lvar.prohibit_numbered_param;
-            if prohibit.is_none() {
-                *prohibit = Some(loc);
-            }
-            match c.kind {
-                ScopeKind::Block => {} //outer += 1,
-                ScopeKind::For => {}
-                _ => break,
-            }
-        }
-        self.scope.last_mut().unwrap().lvar.numbered_param = Some(loc);
-        /*if let Some(a) = &self.extern_context {
-            return a.find_lvar(id);
-        }*/
-        Ok(())
-    }
-
-    /// Peek next token (skipping line terminators).
-    fn peek(&mut self) -> Result<Token, LexerErr> {
-        self.lexer.peek_token_skip_lt()
-    }
-
-    /// Peek next token (no skipping line terminators).
-    fn peek_no_term(&mut self) -> Result<Token, LexerErr> {
-        self.lexer.peek_token()
-    }
-
-    /// Peek next token (no skipping line terminators), and check whether the token is `punct` or not.
-    fn peek_punct_no_term(&mut self, punct: Punct) -> bool {
-        match self.lexer.peek_token() {
-            Ok(tok) => tok.kind == TokenKind::Punct(punct),
-            Err(_) => false,
-        }
-    }
-
-    /// Examine the next token, and return true if it is a line terminator.
-    fn is_line_term(&mut self) -> Result<bool, LexerErr> {
-        Ok(self.peek_no_term()?.is_line_term())
-    }
-
-    fn loc(&mut self) -> Loc {
-        self.peek_no_term().unwrap().loc()
-    }
-
-    fn prev_loc(&self) -> Loc {
-        self.prev_loc
-    }
-
-    /// Get next token (skipping line terminators).
-    /// Return RubyError if it was EOF.
-    fn get(&mut self) -> Result<Token, LexerErr> {
-        loop {
-            let tok = self.lexer.get_token(false)?;
-            if tok.is_eof() {
-                return Err(error_eof(tok.loc()));
-            }
-            if !tok.is_line_term() {
-                self.prev_loc = tok.loc;
-                return Ok(tok);
-            }
-        }
-    }
-
-    /// Get next token (skipping line terminators).
-    /// Do not allow assignment operators.
-    /// Return RubyError if it was EOF.
-    fn get_primary(&mut self) -> Result<Token, LexerErr> {
-        loop {
-            let tok = self.lexer.get_token(true)?;
-            if tok.is_eof() {
-                return Err(error_eof(tok.loc()));
-            }
-            if !tok.is_line_term() {
-                self.prev_loc = tok.loc;
-                return Ok(tok);
-            }
-        }
-    }
-
-    /// Get next token (no skipping line terminators).
-    fn get_no_skip_line_term(&mut self) -> Result<Token, LexerErr> {
-        let tok = self.lexer.get_token(false)?;
-        self.prev_loc = tok.loc;
-        Ok(tok)
-    }
-
-    fn consume_char(&mut self, ch: char) -> bool {
-        self.lexer.consume(ch)
-    }
-
-    /// If the next token is Ident, consume and return Some(it).
-    /// If not, return None.
-    fn consume_ident(&mut self) -> Result<Option<String>, LexerErr> {
-        match self.peek()?.kind {
-            TokenKind::Ident(s) => {
-                self.get()?;
-                Ok(Some(s))
-            }
-            TokenKind::NumberedParam(i, _) => Err(error_numbered_param(self.prev_loc(), i)),
-            _ => Ok(None),
-        }
-    }
-
-    /// If the next token is an expected kind of Punctuator, get it and return true.
-    /// Otherwise, return false.
-    fn consume_punct(&mut self, expect: Punct) -> Result<bool, LexerErr> {
-        match self.peek()?.kind {
-            TokenKind::Punct(punct) if punct == expect => {
-                self.get()?;
-                Ok(true)
-            }
-            _ => Ok(false),
-        }
-    }
-
-    fn consume_punct_no_term(&mut self, expect: Punct) -> Result<bool, LexerErr> {
-        if TokenKind::Punct(expect) == self.peek_no_term()?.kind {
-            self.get()?;
-            Ok(true)
-        } else {
-            Ok(false)
-        }
-    }
-
-    fn consume_assign_op_no_term(&mut self) -> Result<Option<BinOp>, LexerErr> {
-        if let TokenKind::Punct(Punct::AssignOp(op)) = self.peek_no_term()?.kind {
-            Ok(Some(op))
-        } else {
-            Ok(None)
-        }
-    }
-
-    /// If next token is an expected kind of Reserved keyeord, get it and return true.
-    /// Otherwise, return false.
-    fn consume_reserved(&mut self, expect: Reserved) -> Result<bool, LexerErr> {
-        match self.peek()?.kind {
-            TokenKind::Reserved(reserved) if reserved == expect => {
-                self.get()?;
-                Ok(true)
-            }
-            _ => Ok(false),
-        }
-    }
-
-    fn consume_reserved_no_skip_line_term(&mut self, expect: Reserved) -> Result<bool, LexerErr> {
-        if TokenKind::Reserved(expect) == self.peek_no_term()?.kind {
-            self.get()?;
-            Ok(true)
-        } else {
-            Ok(false)
-        }
-    }
-
-    /// Get the next token if it is a line terminator or ';' or EOF, and return true,
-    /// Otherwise, return false.
-    fn consume_term(&mut self) -> Result<bool, LexerErr> {
-        if !self.peek_no_term()?.is_term() {
-            return Ok(false);
-        };
-        while self.peek_no_term()?.is_term() {
-            if self.get_no_skip_line_term()?.is_eof() {
-                return Ok(true);
-            }
-        }
-        Ok(true)
-    }
-
-    /// Get the next token and examine whether it is an expected Reserved.
-    /// If not, return RubyError.
-    fn expect_reserved(&mut self, expect: Reserved) -> Result<(), LexerErr> {
-        match &self.get()?.kind {
-            TokenKind::Reserved(reserved) if *reserved == expect => Ok(()),
-            t => Err(error_unexpected(
-                self.prev_loc(),
-                format!("Expect {:?} Got {:?}", expect, t),
-            )),
-        }
-    }
-
-    /// Get the next token and examine whether it is an expected Punct.
-    /// If not, return RubyError.
-    fn expect_punct(&mut self, expect: Punct) -> Result<(), LexerErr> {
-        match &self.get()?.kind {
-            TokenKind::Punct(punct) if *punct == expect => Ok(()),
-            t => Err(error_unexpected(
-                self.prev_loc(),
-                format!("Expect {:?} Got {:?}", expect, t),
-            )),
-        }
-    }
-
-    /// Get the next token and examine whether it is Ident.
-    /// Return IdentId of the Ident.
-    /// If not, return RubyError.
-    fn expect_ident(&mut self) -> Result<String, LexerErr> {
-        match self.get()?.kind {
-            TokenKind::Ident(name) => Ok(name),
-            TokenKind::NumberedParam(i, _) => Err(error_numbered_param(self.prev_loc(), i)),
-            _ => Err(error_unexpected(self.prev_loc(), "Expect identifier.")),
-        }
-    }
-
-    /// Get the next token and examine whether it is Const.
-    /// Return IdentId of the Const.
-    /// If not, return RubyError.
-    fn expect_const(&mut self) -> Result<String, LexerErr> {
-        match self.get()?.kind {
-            TokenKind::Const(s) => Ok(s),
-            _ => Err(error_unexpected(self.prev_loc(), "Expect constant.")),
-        }
-    }
-
-    fn read_method_name(&mut self, allow_assign_like: bool) -> Result<(String, Loc), LexerErr> {
-        self.lexer.read_method_name(allow_assign_like)
-    }
-
-    fn read_method_ext(&mut self, s: String) -> Result<String, LexerErr> {
-        self.lexer.read_method_ext(s)
-    }
-}
-
-impl<'a, OuterContext: LocalsContext> Parser<'a, OuterContext> {
-    /// Parse operator which can be defined as a method.
-    /// Return IdentId of the operator.
-    fn parse_op_definable(&mut self, punct: &Punct) -> Result<&'static str, LexerErr> {
-        // TODO: must support
-        // ` !
-        match punct {
-            Punct::Plus => {
-                if self.consume_char('@') {
-                    Ok("+@")
-                } else {
-                    Ok("+")
-                }
-            }
-            Punct::Minus => {
-                if self.consume_char('@') {
-                    Ok("-@")
-                } else {
-                    Ok("-")
-                }
-            }
-            Punct::DMul => Ok("**"),
-            Punct::Mul => Ok("*"),
-            Punct::Div => Ok("/"),
-            Punct::Rem => Ok("%"),
-            Punct::Shl => Ok("<<"),
-            Punct::Shr => Ok(">>"),
-            Punct::BitAnd => Ok("&"),
-            Punct::BitOr => Ok("|"),
-            Punct::BitNot => {
-                // `def ~@` is an alternative spelling of `def ~`.
-                // CRuby accepts both and uses `~` as the method name.
-                self.consume_char('@');
-                Ok("~")
-            }
-
-            Punct::Cmp => Ok("<=>"),
-            Punct::Eq => Ok("=="),
-            Punct::Ne => Ok("!="),
-            Punct::Lt => Ok("<"),
-            Punct::Le => Ok("<="),
-            Punct::Gt => Ok(">"),
-            Punct::Ge => Ok(">="),
-            Punct::TEq => Ok("==="),
-            Punct::Match => Ok("=~"),
-            Punct::BitXor => Ok("^"),
-            Punct::Not => Ok("!"),
-            Punct::Unmatch => Ok("!~"),
-            Punct::LBracket => {
-                if self.consume_punct_no_term(Punct::RBracket)? {
-                    if self.consume_punct_no_term(Punct::Assign)? {
-                        Ok("[]=")
-                    } else {
-                        Ok("[]")
-                    }
-                } else {
-                    let loc = self.loc();
-                    Err(error_unexpected(loc, "Invalid operator."))
-                }
-            }
-            _ => Err(error_unexpected(self.prev_loc(), "Invalid operator.")),
-        }
-    }
-
-    fn parse_then(&mut self) -> Result<(), LexerErr> {
-        if self.consume_term()? {
-            self.consume_reserved(Reserved::Then)?;
-            return Ok(());
-        }
-        self.expect_reserved(Reserved::Then)
-    }
-
-    fn parse_do(&mut self) -> Result<(), LexerErr> {
-        if self.consume_term()? {
-            return Ok(());
-        }
-        self.expect_reserved(Reserved::Do)
-    }
-
-    /// Parse formal parameters.
-    /// required, optional = defaule, *rest, post_required, kw: default, **rest_kw, &block
-    ///
-    /// `is_block` is set by the caller to indicate that the parameter list
-    /// belongs to a block (`|...|`). In that case, a trailing comma after
-    /// a required parameter (`|k,|`) injects a synthetic anonymous post
-    /// parameter — bumping the positional arity to >1 so block auto-splat
-    /// of a single Array argument fires. Method `def` and lambda contexts
-    /// pass `false` and reject trailing commas via the usual SyntaxError
-    /// path of the surrounding parser.
-    fn parse_formal_params(
-        &mut self,
-        terminator: impl Into<Option<Punct>>,
-        is_block: bool,
-    ) -> Result<Vec<FormalParam>, LexerErr> {
-        #[derive(Debug, Clone, PartialEq, PartialOrd)]
-        enum Kind {
-            Required,
-            Optional,
-            Rest,
-            PostReq,
-            KeyWord,
-            KWRest,
-        }
-
-        let terminator = terminator.into();
-        let mut args = vec![];
-        let mut state = Kind::Required;
-
-        loop {
-            if let Some(term) = terminator {
-                if self.consume_punct(term)? {
-                    return Ok(args);
-                }
-            }
-            let mut loc = self.loc();
-            if self.consume_punct(Punct::LParen)? {
-                args.push(self.parse_destruct_param()?);
-            } else if self.consume_punct(Punct::Range3)? {
-                // Argument delegation
-                if state > Kind::Required {
-                    return Err(error_unexpected(
-                        loc,
-                        "parameter forwarding is not allowed in ths position.",
-                    ));
-                }
-                args.push(FormalParam::forwarding(loc));
-                self.new_forwarding_param(loc)?;
-                break;
-            } else if self.consume_punct(Punct::BitAnd)? {
-                // Block param
-                loc = loc.merge(self.prev_loc());
-                match self.consume_ident()? {
-                    Some(name) => {
-                        args.push(FormalParam::block(name.clone(), loc));
-                        self.new_block_param(name, loc)?;
-                    }
-                    None => {
-                        args.push(FormalParam::block_discard(loc));
-                    }
-                }
-                break;
-            } else if self.consume_punct(Punct::Mul)? {
-                // Splat(Rest) param
-                loc = loc.merge(self.prev_loc());
-                if state >= Kind::Rest {
-                    return Err(error_unexpected(
-                        loc,
-                        "Rest parameter is not allowed in ths position.",
-                    ));
-                };
-                state = Kind::Rest;
-                match self.consume_ident()? {
-                    Some(name) => {
-                        self.new_param(&name, self.prev_loc())?;
-                        args.push(FormalParam::rest(name, loc));
-                    }
-                    None => args.push(FormalParam::rest_discard(loc)),
-                }
-            } else if self.consume_punct(Punct::DMul)? {
-                // Keyword rest param
-                loc = loc.merge(self.prev_loc());
-                if state >= Kind::KWRest {
-                    return Err(error_unexpected(
-                        loc,
-                        "Keyword rest parameter is not allowed in ths position.",
-                    ));
-                }
-                state = Kind::KWRest;
-                match self.consume_ident()? {
-                    Some(name) => {
-                        args.push(FormalParam::kwrest(name.clone(), loc));
-                        self.new_kwrest_param(name, self.prev_loc())?;
-                    }
-                    None => args.push(FormalParam::kwrest_discard(loc)),
-                }
-            } else {
-                // Accept a reserved word as a keyword argument name when it
-                // is immediately followed by `:` (e.g. `def foo(if: true)`).
-                // CRuby allows reserved words only for keyword parameters,
-                // never for positional ones, so this branch handles exactly
-                // the `reserved-name:` case and does not allow `Assign`.
-                let reserved_kw = {
-                    let tok = self.peek_no_term()?;
-                    if let TokenKind::Reserved(r) = tok.kind
-                        && self.lexer.has_trailing_colon(&tok)
-                    {
-                        Some((r.as_str().to_string(), tok.loc))
-                    } else {
-                        None
-                    }
-                };
-                let name = if let Some((kw_name, _)) = &reserved_kw {
-                    // Consume the reserved token.
-                    self.get()?;
-                    kw_name.clone()
-                } else {
-                    self.expect_ident()?
-                };
-                if reserved_kw.is_none() && self.consume_punct(Punct::Assign)? {
-                    // Optional param
-                    match state {
-                        Kind::Required => state = Kind::Optional,
-                        Kind::Optional => {}
-                        _ => {
-                            return Err(error_unexpected(
-                                loc,
-                                "Optional parameter is not allowed in ths position.",
-                            ));
-                        }
-                    };
-                    // The parameter is in scope (as `nil`) while its own
-                    // default expression is parsed, matching CRuby:
-                    // `->(a = a) { a }` is `nil`, `->(a = a + 1) {}`
-                    // raises NoMethodError. Register the name *before*
-                    // parsing the default so a self-reference resolves to
-                    // the local var instead of a method call on `self`.
-                    self.new_param(&name, loc)?;
-                    let default = if let Some(Punct::BitOr) = terminator {
-                        let node = self.parse_primary()?;
-                        node
-                    } else {
-                        self.parse_arg(false)?
-                    };
-                    loc = loc.merge(self.prev_loc());
-                    args.push(FormalParam::optional(name, default, loc));
-                } else if self.consume_punct_no_term(Punct::Colon)? {
-                    // Keyword param
-                    let next = self.peek_no_term()?.kind;
-                    let default =
-                        if next == TokenKind::Punct(Punct::Comma) || next == TokenKind::LineTerm {
-                            None
-                        } else if let Some(term) = terminator {
-                            if next == TokenKind::Punct(term) {
-                                None
-                            } else {
-                                Some(self.parse_arg(false)?)
-                            }
-                        } else {
-                            Some(self.parse_arg(false)?)
-                        };
-                    loc = loc.merge(self.prev_loc());
-                    if state == Kind::KWRest {
-                        return Err(error_unexpected(
-                            loc,
-                            "Keyword parameter is not allowed in ths position.",
-                        ));
-                    } else {
-                        state = Kind::KeyWord;
-                    };
-                    let lvar = self.new_param(&name, loc)?;
-                    args.push(FormalParam::keyword(name, default, loc));
-                    self.add_kw_param(lvar);
-                } else {
-                    // Required param
-                    //loc = self.prev_loc();
-                    match state {
-                        Kind::Required => {
-                            self.new_param(&name, loc)?;
-                            args.push(FormalParam::req_param(name, loc));
-                        }
-                        Kind::PostReq | Kind::Optional | Kind::Rest => {
-                            self.new_param(&name, loc)?;
-                            args.push(FormalParam::post(name, loc));
-                            state = Kind::PostReq;
-                        }
-                        _ => {
-                            return Err(error_unexpected(
-                                loc,
-                                "Required parameter is not allowed in ths position.",
-                            ));
-                        }
-                    }
-                };
-            }
-            match terminator {
-                Some(_) => {
-                    if !self.consume_punct(Punct::Comma)? {
-                        break;
-                    }
-                }
-                None => {
-                    if !self.consume_punct_no_term(Punct::Comma)? {
-                        break;
-                    }
-                }
-            }
-            // Trailing comma in a block parameter list (`|k,|`): when the
-            // next token is the terminator, treat the comma as an
-            // auto-splat hint by injecting an anonymous post parameter.
-            // Only valid while we're still parsing required positionals;
-            // `|*a,|` etc. fall through to the duplicate-rest error path
-            // matching CRuby.
-            if is_block
-                && state == Kind::Required
-                && let Some(term) = terminator
-                && self.peek_punct_no_term(term)
-            {
-                let trailing_loc = self.prev_loc();
-                args.push(FormalParam::post_discard(trailing_loc));
-                break;
-            }
-        }
-        if let Some(term) = terminator {
-            self.expect_punct(term)?;
-        }
-        Ok(args)
-    }
-
-    fn parse_destruct_param(&mut self) -> Result<FormalParam, LexerErr> {
-        let loc = self.loc();
-        let name = self.expect_ident()?;
-        self.new_param(&name, loc)?;
-        let mut idents = vec![(name, loc)];
-        while self.consume_punct(Punct::Comma)? {
-            if self.consume_punct(Punct::RParen)? {
-                return Ok(FormalParam::destruct_param(idents));
-            }
-            let loc = self.loc();
-            let name = self.expect_ident()?;
-            self.new_param(&name, loc)?;
-            idents.push((name, loc));
-        }
-        self.expect_punct(Punct::RParen)?;
-        Ok(FormalParam::destruct_param(idents))
-    }
-}
-
-fn error_unexpected(loc: Loc, msg: impl Into<String>) -> LexerErr {
-    LexerErr(ParseErrKind::SyntaxError(msg.into()), loc)
-}
-
-fn error_numbered_param(loc: Loc, i: u8) -> LexerErr {
-    LexerErr(
-        ParseErrKind::SyntaxError(format!("_{i} is reserved for numbered parameter")),
-        loc,
-    )
-}
-
-fn error_nested_numbered_param(loc: Loc) -> LexerErr {
-    LexerErr(
-        ParseErrKind::SyntaxError("numbered parameter is already used".to_string()),
-        loc,
-    )
-}
-
-fn error_eof(loc: Loc) -> LexerErr {
-    LexerErr(ParseErrKind::UnexpectedEOF, loc)
-}
-
-fn parse(
-    code: String,
-    path: PathBuf,
-    extern_context: Option<&impl LocalsContext>,
-    parse_context: LvarScope,
-    line_offset: i64,
-) -> Result<ParseResult, ParseErr> {
-    match Parser::parse(&code, path.clone(), extern_context, parse_context, line_offset) {
-        Ok((node, lvar_collector, tok)) => {
-            let source_info = SourceInfoRef::new(SourceInfo::new_eval(path, code, line_offset));
-            if tok.is_eof() {
-                let result = ParseResult {
-                    node,
-                    lvar_collector,
-                    source_info,
-                };
-                Ok(result)
-            } else {
-                let err = error_unexpected(tok.loc(), "Expected end-of-input.");
-                Err(ParseErr::from_lexer_err(err, source_info))
-            }
-        }
-        Err(err) => {
-            let source_info = SourceInfoRef::new(SourceInfo::new_eval(path, code, line_offset));
-            Err(ParseErr::from_lexer_err(err, source_info))
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct ParseResult {
-    pub node: Node,
-    pub lvar_collector: LvarCollector,
-    pub source_info: SourceInfoRef,
-    //pub id_store: IdentifierTable,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-enum LoopKind {
-    Top,
-    Block,
-    While,
-    For,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-enum ScopeKind {
-    Eval,
-    Class,
-    Method,
-    Block,
-    For,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-struct LvarScope {
-    kind: ScopeKind,
-    lvar: LvarCollector,
-}
-
-impl LvarScope {
-    fn new_method() -> Self {
-        LvarScope {
-            lvar: LvarCollector::new(),
-            kind: ScopeKind::Method,
-        }
-    }
-
-    fn new_eval(lvar_collector: Option<LvarCollector>) -> Self {
-        LvarScope {
-            lvar: lvar_collector.unwrap_or_default(),
-            kind: ScopeKind::Eval,
-        }
-    }
-
-    fn new_class(lvar_collector: Option<LvarCollector>) -> Self {
-        LvarScope {
-            lvar: lvar_collector.unwrap_or_default(),
-            kind: ScopeKind::Class,
-        }
-    }
-
-    fn new_block(lvar_collector: Option<LvarCollector>) -> Self {
-        LvarScope {
-            lvar: lvar_collector.unwrap_or_default(),
-            kind: ScopeKind::Block,
-        }
-    }
-
-    fn new_for() -> Self {
-        LvarScope {
-            lvar: LvarCollector::new(),
-            kind: ScopeKind::For,
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct RescueEntry {
-    /// The exception classes for this rescue clause.
-    pub exception_list: Vec<Node>,
-    /// Assignment destination for error value in rescue clause.
-    pub assign: Option<Box<Node>>,
-    /// The body of this rescue clause.
-    pub body: Box<Node>,
-}
-
-impl RescueEntry {
-    fn new(exception_list: Vec<Node>, assign: Option<Node>, body: Node) -> Self {
-        Self {
-            exception_list,
-            assign: assign.map(Box::new),
-            body: Box::new(body),
-        }
-    }
-
-    pub(crate) fn new_postfix(body: Node) -> Self {
-        Self {
-            exception_list: vec![],
-            assign: None,
-            body: Box::new(body),
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum NReal {
-    Integer(i64),
-    Bignum(BigInt),
-    Float(f64),
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    fn parse_test(code: &str) {
-        let node = Parser::<DummyContext>::parse(
-            code,
-            std::path::PathBuf::new(),
-            None,
-            LvarScope::new_eval(None),
-            0,
-        )
-        .unwrap()
-        .0;
-        dbg!(node);
-    }
-
-    fn parse_node(code: &str, expected: Node) {
-        let node = Parser::<DummyContext>::parse(
-            code,
-            std::path::PathBuf::new(),
-            None,
-            LvarScope::new_eval(None),
-            0,
-        )
-        .unwrap()
-        .0;
-        assert_eq!(node, expected);
-    }
-
-    fn parse_test_err(code: &str) {
-        Parser::<DummyContext>::parse(
-            code,
-            std::path::PathBuf::new(),
-            None,
-            LvarScope::new_eval(None),
-            0,
-        )
-        .unwrap_err();
-    }
-
-    #[test]
-    fn test() {
-        parse_test("1.times {|x,y|}");
-        parse_test("1.times {|x,|}");
-        parse_test("1.times {|x,(y,z)|}");
-        parse_test("1.times {|x,(y,z,),|}");
-        parse_test_err("1.times {|x,(y,z|}");
-        parse_test_err("1.times {|x,(y,z)}");
-
-        parse_test(
-            r#"1.times {|x
-            ,(y,z,),|}"#,
-        );
-        parse_test(
-            r#"1.times {|x
-            ,(y
-                ,z,),|}"#,
-        );
-        parse_test(
-            r#"
-            def f y,z
-            end"#,
-        );
-        parse_test(
-            r#"
-            def f y,
-            z
-            end"#,
-        );
-        parse_test_err(
-            r#"
-            def f y
-            ,z
-            end"#,
-        );
-        parse_test(
-            r#"
-        for i in 0..5
-            redo
-        end
-        "#,
-        );
-    }
-
-    #[test]
-    fn method_call() {
-        parse_test("a()");
-        parse_test("a(100)");
-        parse_test("a(:a=>1, :b=>2)");
-        parse_test("a(:a=>1, :b=>2, )");
-    }
-
-    #[test]
-    fn hash() {
-        parse_test("[:a => 5]");
-        parse_test("[][:a => 5]");
-        parse_test("p(:a => 5)");
-        parse_test("p :a => 5");
-    }
-
-    #[test]
-    fn comment_in_method_chain() {
-        parse_test(
-            r#"
-            [1,2,3].
-              # comment
-              map {|x| x * 2}
-            "#,
-        );
-        parse_test(
-            r#"
-            a. # comment
-              b. # another comment
-              c
-            "#,
-        );
-    }
-
-    #[test]
-    fn hash_splat() {
-        parse_test("{**h}");
-        parse_test("{a: 1, **h}");
-        parse_test("{a: 1, **h, b: 2}");
-        parse_test("{**h1, **h2}");
-        parse_test("{1 => 2, **h}");
-    }
-
-    #[test]
-    fn special_vars() {
-        parse_node(
-            "$&",
-            Annot {
-                kind: NodeKind::GlobalVar("$&".to_string()),
-                loc: Loc(0, 1),
-            },
-        );
-        parse_node(
-            "$'",
-            Annot {
-                kind: NodeKind::GlobalVar("$'".to_string()),
-                loc: Loc(0, 1),
-            },
-        );
-        parse_node(
-            "$LOAD_PATH",
-            Annot {
-                kind: NodeKind::GlobalVar("$LOAD_PATH".to_string()),
-                loc: Loc(0, 9),
-            },
-        );
-        parse_node(
-            "$:",
-            Annot {
-                kind: NodeKind::GlobalVar("$:".to_string()),
-                loc: Loc(0, 1),
-            },
-        );
-        parse_node(
-            "$LOADED_FEATURES",
-            Annot {
-                kind: NodeKind::GlobalVar("$LOADED_FEATURES".to_string()),
-                loc: Loc(0, 15),
-            },
-        );
-        parse_node(
-            "$\"",
-            Annot {
-                kind: NodeKind::GlobalVar("$\"".to_string()),
-                loc: Loc(0, 1),
-            },
-        );
-    }
-
-    #[test]
-    fn test_err() {
-        parse_test_err(
-            r#"
-        def f
-          class C; end
-        end
-        "#,
-        );
-        parse_test_err(
-            r#"
-        def f
-          1.times do
-            class C; end
-          end
-        end
-        "#,
-        );
-        //parse_test_err("break");
-        //parse_test_err("next");
-        //parse_test_err("redo");
-        parse_test_err(
-            r#"
-        for i in 0..5
-            redo 5
-        end
-        "#,
-        );
-        parse_test("defined? break");
-        parse_test("defined? next");
-        parse_test("defined? redo");
-        /*parse_test_err(
-            r#"
-        def f
-          break
-        end
-        "#,
-        );
-        parse_test_err(
-            r#"
-        def f
-          next
-        end
-        "#,
-        );*/
-    }
-
-    #[test]
-    fn numbered_param_basic() {
-        // Basic usage in block
-        parse_test("[1,2,3].map { _1 }");
-        parse_test("[1,2,3].each { _1 + _2 }");
-        parse_test("[1,2,3].map { _1 * 2 }");
-        // Multiple numbered params
-        parse_test("[[1,2],[3,4]].map { _1 + _2 }");
-        // In do..end block
-        parse_test("[1,2,3].each do _1 end");
-        parse_test("[1,2,3].each do _1 + _2 end");
-        // Nested expressions with numbered params
-        parse_test("[1,2,3].map { _1.to_s }");
-        parse_test("[1,2,3].map { _1 > 1 ? _1 : 0 }");
-    }
-
-    #[test]
-    fn numbered_param_as_func_call() {
-        // _1() should be parsed as a method call (parse_func_call)
-        parse_test("1.times { _1() }");
-        parse_test("1.times { _1(42) }");
-        parse_test("1.times { _2(1, 2) }");
-    }
-
-    #[test]
-    fn numbered_param_errors() {
-        // Numbered param cannot be used in comma expression (multiple assignment lhs)
-        parse_test_err("1.times { _1, x = 1, 2 }");
-        // Numbered param in def method name position
-        parse_test_err("def _1; end");
-        // Numbered param used as formal param name
-        parse_test_err("1.times { |_1| }");
-    }
-
-    #[test]
-    fn parse_func_call() {
-        // Simple function calls
-        parse_test("foo()");
-        parse_test("foo(1)");
-        parse_test("foo(1, 2, 3)");
-        // Function call with block
-        parse_test("foo {}");
-        parse_test("foo do end");
-        parse_test("foo { |x| x }");
-        parse_test("foo do |x| x end");
-        // Function call with args and block
-        parse_test("foo(1) {}");
-        parse_test("foo(1) do end");
-        parse_test("foo(1, 2) { |x| x }");
-        // Function call with keyword args
-        parse_test("foo(a: 1, b: 2)");
-        parse_test("foo(1, a: 1)");
-        // No-paren function calls
-        parse_test("foo 1");
-        parse_test("foo 1, 2");
-        parse_test("foo 1, a: 2");
-        // Splat in args
-        parse_test("foo(*a)");
-        parse_test("foo(1, *a)");
-        // Block arg
-        parse_test("foo(&b)");
-        parse_test("foo(1, &b)");
-    }
-
-    #[test]
-    fn parse_def_operators() {
-        parse_test("class Foo; def **(other); end; end");
-        //parse_test("class Foo; def !; end; end");
-        parse_test("class Foo; def !~(other); end; end");
-    }
-
-    #[test]
-    fn parse_alias_special_var() {
-        parse_test("alias $MATCH $&");
-        parse_test("alias $PREMATCH $'");
-        parse_test("alias $LOAD_PATH $:");
-    }
-
-    #[test]
-    fn parse_alias_method() {
-        parse_test("class Foo; def bar; end; alias baz bar; end");
-    }
-}
