--- ruruby-parse/src/parser/arguments.rs
@@ -1,260 +1,0 @@
-use super::*;
-
-impl<'a, OuterContext: LocalsContext> Parser<'a, OuterContext> {
-    ///
-    /// Parse arguments with parentheses.
-    ///
-    /// - f(arg0, arg1, ...)
-    pub(super) fn parse_arguments_paren(
-        &mut self,
-        allow_block: bool,
-    ) -> Result<Option<ArgList>, LexerErr> {
-        if self.lexer.trailing_lparen() {
-            assert!(self.consume_punct_no_term(Punct::LParen).unwrap());
-            let arglist = self.parse_arglist_block(Punct::RParen, allow_block)?;
-            Ok(Some(arglist))
-        } else {
-            Ok(None)
-        }
-    }
-
-    ///
-    /// Parse arguments without parentheses.
-    ///
-    /// - f arg0, arg1, ...
-    pub(super) fn parse_arguments_unparen(
-        &mut self,
-        allow_block: bool,
-    ) -> Result<Option<ArgList>, LexerErr> {
-        if self.is_command() {
-            let arglist = self.parse_arglist_block(None, allow_block)?;
-            Ok(Some(arglist))
-        } else {
-            Ok(None)
-        }
-    }
-
-    fn parse_arglist_block(
-        &mut self,
-        delimiter: impl Into<Option<Punct>>,
-        allow_block: bool,
-    ) -> Result<ArgList, LexerErr> {
-        let mut arglist = self.parse_arglist(delimiter)?;
-        if allow_block && let Some(actual_block) = self.parse_block()? {
-            if arglist.block.is_some() {
-                return Err(error_unexpected(
-                    actual_block.loc(),
-                    "Both block arg and actual block given.",
-                ));
-            }
-            arglist.block = Some(actual_block);
-        };
-        Ok(arglist)
-    }
-
-    /// Parse argument list.
-    /// arg, *splat_arg, kw: kw_arg, **double_splat_arg, &block <punct>
-    /// punct: punctuator for terminating arg list. Set None for unparenthesized argument list.
-    fn parse_arglist(&mut self, punct: impl Into<Option<Punct>>) -> Result<ArgList, LexerErr> {
-        let punct = punct.into();
-        let mut arglist = ArgList::default();
-        if self.peek()?.kind == TokenKind::Punct(Punct::Comma) {
-            return Ok(arglist);
-        }
-        let mut as_hash_splat = vec![];
-        let mut kw_args = vec![];
-        // `closed` tracks whether the terminating punct has already been
-        // consumed inside the loop (happens for empty arglists and for a
-        // trailing comma immediately before the closing punct).  When set,
-        // the post-loop code skips the redundant punct consume but still
-        // runs the kw_args / hash_splat finalization below.
-        let mut closed = false;
-        loop {
-            if let Some(punct) = punct {
-                if self.consume_punct(punct)? {
-                    closed = true;
-                    break;
-                }
-            }
-            if self.consume_punct(Punct::Range3)? {
-                let next = self.peek()?;
-                if next.kind == TokenKind::Punct(Punct::RParen)
-                    || next.kind == TokenKind::Punct(Punct::Comma)
-                    || next.is_term()
-                {
-                    // argument forwarding: foo(...)
-                    self.check_forwarding()?;
-                    arglist.forwarding = true;
-                } else {
-                    // beginless exclusive range: foo(...expr)
-                    let loc = self.prev_loc();
-                    let rhs = self.parse_arg_logical_or()?;
-                    let range_loc = loc.merge(rhs.loc());
-                    let node = Node::new_range(None, Some(rhs), true, range_loc);
-                    arglist.args.push(node);
-                }
-            } else if self.consume_punct(Punct::Mul)? {
-                // splat argument
-                let loc = self.prev_loc();
-                let array = self.parse_arg(false)?;
-                arglist.splat = true;
-                arglist.args.push(Node::new_splat(array, loc));
-            } else if self.consume_punct(Punct::DMul)? {
-                // double splat argument
-                arglist.hash_splat.push(self.parse_arg(false)?);
-            } else if self.consume_punct(Punct::BitAnd)? {
-                // block argument
-                let next = self.peek()?;
-                if next.kind == TokenKind::Punct(Punct::RParen)
-                    || next.kind == TokenKind::Punct(Punct::Comma)
-                    || next.is_term()
-                {
-                    // anonymous block forwarding: foo(&)
-                    arglist.delegate_block = true;
-                } else {
-                    arglist.block = Some(Box::new(self.parse_arg(false)?));
-                }
-            } else {
-                let node = self.parse_arg(false)?;
-                if self.consume_punct(Punct::FatArrow)? {
-                    let value = self.parse_arg(false)?;
-                    if as_hash_splat.is_empty()
-                        && let NodeKind::Symbol(sym) = &node.kind
-                    {
-                        // keyword args
-                        kw_args.push((sym.to_string(), value));
-                    } else {
-                        if as_hash_splat.is_empty() {
-                            as_hash_splat.extend(
-                                std::mem::take(&mut kw_args)
-                                    .into_iter()
-                                    .map(|(id, value)| (Node::new_symbol(id, node.loc()), value)),
-                            );
-                        }
-                        as_hash_splat.push((node, value));
-                    }
-                } else {
-                    if let Some(id) = node.kind.is_identifier()
-                        && self.consume_punct_no_term(Punct::Colon)?
-                    {
-                        // keyword args
-                        // Check for value-omitted keyword argument shorthand (Ruby 3.1+)
-                        // e.g., foo(x:) is equivalent to foo(x: x).
-                        //
-                        // Inside `(...)`, line terminators are not significant,
-                        // so use `peek` (which skips LineTerm) rather than
-                        // `peek_no_term` — otherwise `foo(\n  x:,\n  y:\n)`
-                        // fails to recognise the shorthand because the token
-                        // immediately after the colon is a newline.
-                        let next_kind = if punct.is_some() {
-                            self.peek()?.kind
-                        } else {
-                            self.peek_no_term()?.kind
-                        };
-                        let is_shorthand = matches!(
-                            next_kind,
-                            TokenKind::Punct(Punct::Comma)
-                                | TokenKind::Punct(Punct::RParen)
-                                | TokenKind::Punct(Punct::RBrace)
-                                | TokenKind::Punct(Punct::RBracket)
-                        );
-                        let value = if is_shorthand {
-                            // Value-omitted shorthand: use the identifier as the value
-                            node.clone()
-                        } else {
-                            self.parse_arg(false)?
-                        };
-                        if as_hash_splat.is_empty() {
-                            kw_args.push((id.to_string(), value));
-                        } else {
-                            as_hash_splat.push((Node::new_symbol(id.to_string(), node.loc), value));
-                        }
-                    } else {
-                        arglist.args.push(node);
-                    }
-                }
-            }
-            if !self.consume_punct(Punct::Comma)? {
-                break;
-            } else {
-                let loc = self.prev_loc();
-                if arglist.block.is_some() {
-                    return Err(error_unexpected(loc, "unexpected ','."));
-                };
-            }
-        }
-
-        if !as_hash_splat.is_empty() {
-            arglist
-                .hash_splat
-                .push(Node::new_hash(as_hash_splat, self.prev_loc()));
-        }
-        if !kw_args.is_empty() {
-            arglist.kw_args = kw_args;
-        }
-
-        if !closed && let Some(punct) = punct {
-            self.consume_punct(punct)?;
-        };
-        Ok(arglist)
-    }
-
-    fn is_command(&mut self) -> bool {
-        let tok = match self.peek_no_term() {
-            Ok(tok) => tok,
-            _ => return false,
-        };
-        if self.lexer.trailing_space() {
-            match tok.kind {
-                TokenKind::LineTerm | TokenKind::Eof => false,
-                TokenKind::Punct(p) => match p {
-                    // Unambiguous argument-starting punctuators: `(`, `[`,
-                    // `::`, `->`, `!`, and `~`. `~` is only ever a unary
-                    // operator, so it always begins an argument (matches
-                    // CRuby's `foo ~x` parse).
-                    Punct::LParen
-                    | Punct::LBracket
-                    | Punct::Scope
-                    | Punct::Arrow
-                    | Punct::Not
-                    | Punct::BitNot => true,
-                    Punct::Colon
-                    | Punct::Plus
-                    | Punct::Minus
-                    | Punct::Mul
-                    | Punct::Div
-                    | Punct::BitAnd
-                    | Punct::Rem
-                    | Punct::Shl => !self.lexer.has_trailing_space(&tok),
-                    _ => false,
-                },
-                TokenKind::Reserved(r) => !matches!(
-                    r,
-                    Reserved::Do
-                        | Reserved::If
-                        | Reserved::Unless
-                        | Reserved::Rescue
-                        | Reserved::While
-                        | Reserved::Until
-                        | Reserved::And
-                        | Reserved::Or
-                        | Reserved::Then
-                        | Reserved::Else
-                        | Reserved::Elsif
-                        | Reserved::End
-                ),
-                _ => true,
-            }
-        } else {
-            matches!(
-                tok.kind,
-                TokenKind::GlobalVar(_)
-                    | TokenKind::InstanceVar(_)
-                    | TokenKind::StringLit(_)
-                    | TokenKind::FloatLit(_)
-                    | TokenKind::BignumLit(_)
-                    | TokenKind::IntegerLit(_)
-            )
-        }
-    }
-}
